cluster/applications/bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Kai Aihara <kaidotio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"

uvicorn = "^0.22.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.4"
httpx = "^0.25.1"
fastapi = "^0.100.0"

slack-bolt = "^1.17.0"

openai = "^1.33.0"
tiktoken = "^0.7.0"
python-magic = "^0.4.27"
playwright = "^1.32.1"
redis = "^4.6.0"
redis-om = "^0.2.1"
boto3 = "^1.26.162"
duckduckgo-search = "^3.9.6"

pydantic = "^1.10.10"

opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-aiohttp-client = "^0.45b0"
opentelemetry-instrumentation-httpx = "^0.45b0"
opentelemetry-instrumentation-requests = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-botocore = "^0.45b0"
opentelemetry-instrumentation-redis = "^0.45b0"
opentelemetry-exporter-prometheus = "^0.45b0"
prometheus-client = "^0.17.0"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"

embedding-retrieval = { path = "../embedding-retrieval", develop = true }
cortex = { path = "../packages/cortex", develop = true }
mistune = "^3.0.2"
google-auth-oauthlib = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
